按照客户要求修改入参结构体


Match-id-bde91f0d37b572140383ff57d6c27167de777d56

+++ src/sdk/mirror_client.cpp
@@ -144,9 +144,9 @@ std::vector<uint16_t> MirrorClient::ListLocalAffinityPt()
     return ans;
 }
 
-uint16_t MirrorClient::SelectingPt(uint32_t objectId, AffinityStrategy affinity)
+uint16_t MirrorClient::SelectingPt(uint64_t objectId, AffinityStrategy affinity)
 {
-    size_t v = std::hash<std::uint32_t>{}(objectId);
+    size_t v = std::hash<std::uint64_t>{}(objectId);
     uint16_t ptId = UINT16_MAX;
     if (LIKELY(affinity == LOCAL_AFFINITY)) {
         std::vector<uint16_t> ptVec = ListLocalAffinityPt();
